#25 Refactor: 패키지 구조를 재정의한다
