+++ src/main/java/hanieum/conik/domain/project/entity/Project.java
@@ -1,8 +1,8 @@
-package hanieum.conik.project.domain.entity;
+package hanieum.conik.domain.project.entity;
 
-import hanieum.conik.project.domain.enumerate.ProjectStatus;
-import hanieum.conik.project.domain.enumerate.ProjectType;
-import hanieum.conik.proposal.domain.entity.AbstractEntity;
+import hanieum.conik.domain.project.enumerate.ProjectStatus;
+import hanieum.conik.domain.project.enumerate.ProjectType;
+import hanieum.conik.domain.proposal.domain.entity.AbstractEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
