+++ src/main/java/hanieum/conik/domain/proposal/domain/entity/AbstractEntity.java
@@ -1,4 +1,4 @@
-package hanieum.conik.proposal.domain.entity;
+package hanieum.conik.domain.proposal.domain.entity;
 
 import hanieum.conik.global.domain.BaseEntity;
 import jakarta.persistence.GeneratedValue;
