+++ src/main/java/hanieum/conik/domain/user/exception/UserException.java
@@ -1,4 +1,4 @@
-package hanieum.conik.user.domain.exception;
+package hanieum.conik.domain.user.exception;
 
 import hanieum.conik.global.apiPayload.exception.ErrorType;
 import hanieum.conik.global.apiPayload.exception.GlobalException;
