+++ src/main/resources/META-INF/Project-orm.xml
@@ -5,7 +5,7 @@
                                      https://hibernate.org/xsd/orm/mapping/mapping-3.1.0.xsd">
     <access>FIELD</access>
 
-    <entity class="hanieum.conik.project.domain.entity.Project" access="FIELD">
+    <entity class="hanieum.conik.domain.project.entity.Project" access="FIELD">
         <table name="project"/>
         <attributes>
             <basic name="userId">
