+++ src/main/resources/META-INF/Proposal-orm.xml
@@ -5,7 +5,7 @@
                                      https://hibernate.org/xsd/orm/mapping/mapping-3.1.0.xsd">
     <access>FIELD</access>
 
-    <entity class="hanieum.conik.proposal.domain.entity.Proposal" access="FIELD">
+    <entity class="hanieum.conik.domain.proposal.domain.entity.Proposal" access="FIELD">
         <table name="proposal"/>
         <attributes>
             <basic name="projectId">
